Hack to enable SQLDescribeParam (and removed need for calling execute)
The patch is not really meant to go into production. Just a small hack
so that I can get on with my library which depends upon HDBC.

The patch adds the function getQueryInfo which given a SQL expression
return the types of the parameters and the type of the result. This
is all done without calling SQLExecute, just SQLPrepare.

If you decide to implement my feature requests, this patch may be of
aid to you.